ADOP-1949: Check and send orders
